+++ src/main/java/net/detalk/api/service/DetalkInitializer.java
@@ -9,6 +9,7 @@ import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import net.detalk.api.domain.Role;
+import net.detalk.api.plan.service.PricingPlanCache;
 import net.detalk.api.repository.RoleRepository;
 import net.detalk.api.support.EnvironmentHolder;
 import net.detalk.api.support.security.SecurityRole;
