+++ src/main/java/net/detalk/api/service/ProductPostService.java
@@ -10,6 +10,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import net.detalk.api.controller.v1.request.UpdateProductPostRequest;
 import net.detalk.api.controller.v1.response.GetProductPostResponse;
+import net.detalk.api.plan.service.PricingPlanService;
 import net.detalk.api.product.domain.ProductLink;
 import net.detalk.api.domain.exception.DuplicateCreatePostException;
 import net.detalk.api.domain.exception.InvalidPageSizeException;
@@ -24,7 +25,7 @@ import net.detalk.api.repository.ProductPostLinkRepository;
 import net.detalk.api.support.CursorPageData;
 import net.detalk.api.product.domain.ProductMaker;
 import net.detalk.api.domain.ProductPostSnapshotAttachmentFile;
-import net.detalk.api.domain.PricingPlan;
+import net.detalk.api.plan.domain.PricingPlan;
 import net.detalk.api.product.domain.Product;
 import net.detalk.api.controller.v1.request.CreateProductPostRequest;
 import net.detalk.api.domain.ProductPost;
