+++ src/test/java/net/detalk/api/service/ProductPostServiceTest.java
@@ -17,7 +17,8 @@ import net.detalk.api.controller.v1.request.CreateProductPostRequest;
 import net.detalk.api.controller.v1.request.UpdateProductPostRequest;
 import net.detalk.api.controller.v1.response.GetProductPostResponse;
 import net.detalk.api.controller.v1.response.GetProductPostResponse.Media;
-import net.detalk.api.domain.PricingPlan;
+import net.detalk.api.plan.domain.PricingPlan;
+import net.detalk.api.plan.service.PricingPlanService;
 import net.detalk.api.product.domain.Product;
 import net.detalk.api.product.domain.ProductLink;
 import net.detalk.api.domain.ProductPost;
